pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "allainclair", email = "allainclair@gmail.com"},
]
dependencies = [
    "httpx>=0.24.1",
    "pydantic>=1.10.9",
    "pyhumps>=3.8.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.3.2",
    "icecream>=2.1.3",
    "ipython>=8.14.0",
    "respx>=0.20.1",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.4.1",
]

[tool.pdm.scripts]
pinterest = "python -m app.pinterest"
